+++ storage-resize-images/CHANGELOG.md
@@ -1,3 +1,10 @@
+## Version 0.3.1
+
+chore - bump dependencies
+fix - add missing recordStartEvent call (#2546)
+feat - add new onStartResize event
+chore(storage-resize-images) - Update model to gemini-2.5-flash
+
 ## Version 0.3.0
 
 fix! - remove backfill, due to architectural flaws.